Use attribute macro for scenario
