+++ Cargo.toml
@@ -166,6 +166,7 @@ regex = "1.6.0"
 repository = { path = "crates/repository" }
 reqwest = { version = "0.11.23", features = ["blocking"] }
 rustfft = "6.0.1"
+scenario = { path = "crates/scenario" }
 semver = "1.0.20"
 serde = { version = "1.0.195", features = ["derive", "rc"] }
 serde-transcode = "1.1.1"
