+++ crates/hulk_behavior_simulator/src/bin/replacement_keeper_test.rs
@@ -2,21 +2,22 @@ use std::time::Duration;
 
 use bevy::prelude::*;
 
+use scenario::scenario;
 use spl_network_messages::{GameState, Penalty, PlayerNumber};
 
 use hulk_behavior_simulator::{
     autoref::{AutorefState, GoalMode},
     game_controller::GameControllerCommand,
     robot::Robot,
-    scenario,
     time::{Ticks, TicksTime},
 };
 use types::roles::Role;
 
-scenario!(replacement_keeper_test, |app: &mut App| {
+#[scenario]
+fn golden_goal(app: &mut App) {
     app.add_systems(Startup, startup);
     app.add_systems(Update, update);
-});
+}
 
 fn startup(
     mut commands: Commands,
